[loader] Add loaded assemblies to domain's list even in AOT
With MonoImageStorage, we shift the responsibility for not loading an image
more than once from a global MonoImage cache to a per-domain MonoAssembly cache
and a global MonoImageStorage cache.  The MonoImageStorage cache allows the
same underlying filesystem object to be opened in different domains.  The
per-domain MonoAssembly cache ensures that we don't load the same assembly more
than once.

When AOTing, there's still a root domain, and we use it to prevent repeatedly
loading assemblies each time they're referenced.