[loader] Don't reuse MonoImages in mono_image_open_a_lot
Reuse the underlying MonoImageStorage, but return fresh MonoImages every time
this function is called.

To prevent loading the same assembly more than once in a domain or in an
AssemblyLoadContext, we rely on MonoAssembly caching and the firing of the
assembly search hook (mono_assembly_invoke_search_hook_internal - the appdomain
hook is mono_domain_assembly_search).

With this change it should be possible for two appdomains to load the same
assembly but resolve its references in two different ways.