XXXX make MonoLoadedImages pass it everywhere
So I think we still need some deduplication of loaded MonoImages - but not
globally.  We want to dedup locally - meaning within a domain (or more
precisely within an assembly load context).  Also when images are actually
netmodules we want to dedup them within the scope of an assembly (any way to
avoid this? it's kind of unnecessary the way that image->files and
image->modules gets filled in).

the problem is that there are just too many ways to get into
mono_image_open_a_lot to give all of them their own de-dup logic.

This kind of doesn't work great for our external API or for embedding uses that
just want our metadata reader and not the execution engine.  I kind of guess
that they want the default MonoLoadedImages of the current domain, but what
even _is_ a domain if nothing is running?

Also this broke finding the mempool of an address in mono_find_image_owner
because we don't have a global list of images anymore.

Also when we close an image we want to remove it from some loaded images list.
So we have to store the MonoLoadedImages* of an image in its fields somewhere.
Right now there's no such field.